refactor(core): move leveldb internal pkg
